requirements.txt
# Veritas News Credibility Analysis System
# Python Package Requirements

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# AI/ML Libraries
torch>=1.13.0
transformers>=4.25.0
numpy>=1.24.0
scipy>=1.10.0

# Web Scraping
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Text Processing
nltk>=3.8.0
regex>=2023.8.8

# Utilities
python-dateutil>=2.8.0
urllib3>=1.26.0
certifi>=2023.7.22

# Development Tools (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0